mv and rename TldrawCanvas
